Move registration to user controller
