+++ app/controllers/application_controller.rb
@@ -4,7 +4,7 @@ class ApplicationController < ActionController::API
 
 
   def current_user
-    User.find_by(id: session[:user_id])
+    @current_user ||= User.find_by(id: session[:user_id])
   end
 
   def logged_in?
