+++ config/initializers/cors.rb
@@ -1,4 +1,4 @@
-# Be sure to restart your server when you modify this file.
+ # Be sure to restart your server when you modify this file.
 
 # Avoid CORS issues when API is called from the frontend app.
 # Handle Cross-Origin Resource Sharing (CORS) in order to accept cross-origin AJAX requests.
